+++ Source/UI/ITab_Pawn_Diary.Expansion.cs
@@ -146,6 +146,7 @@ namespace PawnDiary
 
 
             entryExpansionOverrides[key] = expanded;
+            entryExpansionVersion++;
 
         }